+++ packages/admin/admin/src/EditDialog.routerTabs.test.tsx
@@ -1,6 +1,11 @@
 import { Add, Edit } from "@comet/admin-icons";
+<<<<<<< HEAD
 import { Button, IconButton } from "@mui/material";
 import { DataGrid, type GridSlotsComponent } from "@mui/x-data-grid";
+=======
+import { IconButton } from "@mui/material";
+import { DataGrid } from "@mui/x-data-grid";
+>>>>>>> main
 import { screen, waitFor } from "@testing-library/react";
 import { createMemoryHistory } from "history";
 import { type ReactNode, type RefObject, useRef } from "react";
@@ -8,7 +13,12 @@ import { useIntl } from "react-intl";
 import { Router } from "react-router";
 import { render } from "test-utils";
 
+<<<<<<< HEAD
 import { FillSpace } from "./common/FillSpace";
+=======
+import { Button } from "./common/buttons/Button";
+import { ToolbarActions } from "./common/toolbar/actions/ToolbarActions";
+>>>>>>> main
 import { DataGridToolbar } from "./common/toolbar/DataGridToolbar";
 import { EditDialog } from "./EditDialog";
 import { type IEditDialogApi } from "./EditDialogApiContext";
@@ -111,12 +121,7 @@ describe("EditDialog with Stack, Router Tabs and Grid", () => {
                         slotProps={{
                             toolbar: {
                                 toolbarAction: (
-                                    <Button
-                                        startIcon={<Add />}
-                                        onClick={() => editDialogApi.current?.openAddDialog()}
-                                        variant="contained"
-                                        color="primary"
-                                    >
+                                    <Button startIcon={<Add />} onClick={() => editDialogApi.current?.openAddDialog()}>
                                         Add product
                                     </Button>
                                 ),