+++ packages/admin/admin/src/router/ConfirmationDialog.tsx
@@ -1,4 +1,5 @@
 import { Close, Delete, Save, Warning } from "@comet/admin-icons";
+<<<<<<< HEAD
 import {
     Button,
     type ComponentsOverrides,
@@ -11,10 +12,14 @@ import {
     type Theme,
     Typography,
 } from "@mui/material";
+=======
+import { ComponentsOverrides, Dialog, DialogActions, DialogContent, DialogTitle, IconButton, Theme, Typography } from "@mui/material";
+>>>>>>> main
 import { css, useThemeProps } from "@mui/material/styles";
 import { type ReactNode } from "react";
 import { FormattedMessage } from "react-intl";
 
+import { Button } from "../common/buttons/Button";
 import { createComponentSlot } from "../helpers/createComponentSlot";
 import { type ThemedComponentBaseProps } from "../helpers/ThemedComponentBaseProps";
 import { messages } from "../messages";
@@ -152,13 +157,7 @@ export function RouterConfirmationDialog(inProps: RouterConfirmationDialogProps)
                     <FormattedMessage {...messages.discard} />
                 </DiscardButton>
                 {showSaveButton && (
-                    <SaveButton
-                        startIcon={<Save />}
-                        color="primary"
-                        variant="contained"
-                        onClick={() => handleClose(PromptAction.Save)}
-                        {...slotProps?.saveButton}
-                    >
+                    <SaveButton startIcon={<Save />} onClick={() => handleClose(PromptAction.Save)} {...slotProps?.saveButton}>
                         <FormattedMessage {...messages.save} />
                     </SaveButton>
                 )}