+++ packages/admin/admin/src/translator/BaseTranslationDialog.tsx
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 import {
     Button,
     // eslint-disable-next-line no-restricted-imports
@@ -8,9 +9,13 @@ import {
     Grid,
     Typography,
 } from "@mui/material";
+=======
+import { Dialog, DialogActions, DialogContent, DialogTitle, Grid, Typography } from "@mui/material";
+>>>>>>> main
 import { useState } from "react";
 import { FormattedMessage } from "react-intl";
 
+import { Button } from "../common/buttons/Button";
 import { messages } from "../messages";
 
 interface TranslationDialogBaseProps<T> {
@@ -53,7 +58,7 @@ export const BaseTranslationDialog = <T,>(props: TranslationDialogBaseProps<T>) 
                 </Grid>
             </DialogContent>
             <DialogActions>
-                <Button onClick={onClose} color="primary">
+                <Button variant="textDark" onClick={onClose}>
                     <FormattedMessage {...messages.cancel} />
                 </Button>
                 <Button
@@ -61,8 +66,6 @@ export const BaseTranslationDialog = <T,>(props: TranslationDialogBaseProps<T>) 
                         onApplyTranslation(translation);
                         onClose();
                     }}
-                    color="primary"
-                    variant="contained"
                 >
                     <FormattedMessage {...messages.apply} />
                 </Button>