+++ packages/admin/cms-admin/src/common/header/UserHeaderItem.tsx
@@ -1,7 +1,13 @@
 import { gql, useMutation } from "@apollo/client";
+<<<<<<< HEAD
 import { AppHeaderDropdown, type AppHeaderDropdownProps } from "@comet/admin";
 import { Account, ImpersonateUser, Logout, ThreeDotSaving } from "@comet/admin-icons";
 import { Avatar, AvatarGroup, type AvatarProps, Box, Button, Divider, Link, Typography, useMediaQuery, useTheme } from "@mui/material";
+=======
+import { AppHeaderDropdown, AppHeaderDropdownProps, Button } from "@comet/admin";
+import { Account, ImpersonateUser, Logout, ThreeDotSaving } from "@comet/admin-icons";
+import { Avatar, AvatarGroup, AvatarProps, Box, Divider, Link, Typography, useMediaQuery, useTheme } from "@mui/material";
+>>>>>>> main
 import { css, styled } from "@mui/material/styles";
 import { type PropsWithChildren, type ReactElement, useState } from "react";
 import { FormattedMessage } from "react-intl";
@@ -77,8 +83,6 @@ export function UserHeaderItem(props: PropsWithChildren<UserHeaderItemProps>) {
                 <Box padding={4}>
                     <Button
                         fullWidth
-                        variant="contained"
-                        color="primary"
                         disabled={isSigningOut}
                         startIcon={isSigningOut ? <ThreeDotSaving /> : <Logout />}
                         onClick={async () => {