+++ packages/admin/cms-admin/src/cronJobs/CronJobsGrid.tsx
@@ -1,6 +1,7 @@
 import { gql, useApolloClient, useQuery } from "@apollo/client";
-import { CancelButton, FillSpace, MainContent, StackLink, Toolbar, ToolbarTitleItem, useStackSwitchApi } from "@comet/admin";
+import { Button, CancelButton, FillSpace, MainContent, StackLink, Toolbar, ToolbarTitleItem, useStackSwitchApi } from "@comet/admin";
 import { Play, Time } from "@comet/admin-icons";
+<<<<<<< HEAD
 import {
     Button,
     // eslint-disable-next-line no-restricted-imports
@@ -10,6 +11,9 @@ import {
     DialogTitle,
     IconButton,
 } from "@mui/material";
+=======
+import { Dialog, DialogActions, DialogContent, DialogTitle, IconButton } from "@mui/material";
+>>>>>>> main
 import { DataGrid } from "@mui/x-data-grid";
 import { parseISO } from "date-fns";
 import { useState } from "react";
@@ -177,7 +181,6 @@ export function CronJobsGrid() {
                     <CancelButton onClick={closeDialog} />
 
                     <Button
-                        variant="contained"
                         startIcon={<Play />}
                         onClick={async () => {
                             if (cronJobToStart) {