+++ packages/admin/cms-admin/src/dam/DataGrid/duplicatedFilenames/ManuallyHandleDuplicatedFilenamesDialog.tsx
@@ -1,5 +1,6 @@
-import { Alert } from "@comet/admin";
+import { Alert, Button } from "@comet/admin";
 import { Duplicate, Forward, Save } from "@comet/admin-icons";
+<<<<<<< HEAD
 import {
     Box,
     Button,
@@ -14,6 +15,9 @@ import {
     ListItemText,
     Typography,
 } from "@mui/material";
+=======
+import { Box, Dialog, DialogActions, DialogContent, DialogTitle, Divider, List, ListItem, ListItemText, Typography } from "@mui/material";
+>>>>>>> main
 import { styled } from "@mui/material/styles";
 import { Fragment, type VoidFunctionComponent } from "react";
 import { FormattedMessage } from "react-intl";
@@ -108,7 +112,7 @@ export const ManuallyHandleDuplicatedFilenamesDialog: VoidFunctionComponent<Dupl
             </DialogContent>
             <DialogActions>
                 <Box display="flex" justifyContent="space-between" width="100%">
-                    <Button variant="text" color="secondary" onClick={onSkip} startIcon={<Forward />}>
+                    <Button variant="textDark" onClick={onSkip} startIcon={<Forward />}>
                         <FormattedMessage
                             id="comet.dam.duplicateFilenameDialog.action.skip"
                             defaultMessage="Skip {count, plural, one {file} other {files}}"
@@ -127,7 +131,7 @@ export const ManuallyHandleDuplicatedFilenamesDialog: VoidFunctionComponent<Dupl
                                 }}
                             />
                         </Button>
-                        <Button variant="contained" color="primary" onClick={onUpload} startIcon={<Save />}>
+                        <Button onClick={onUpload} startIcon={<Save />}>
                             <FormattedMessage
                                 id="comet.dam.duplicateFilenameDialog.action.uploadAsCopy"
                                 defaultMessage="Upload as {count, plural, one {copy} other {copies}}"