+++ packages/admin/cms-admin/src/dam/FileActions/ConfirmDeleteDialog.sc.ts
@@ -1,5 +1,10 @@
+import { Button } from "@comet/admin";
 import { Warning } from "@comet/admin-icons";
+<<<<<<< HEAD
 import { DialogContent, type Theme } from "@mui/material";
+=======
+import { DialogContent, Theme } from "@mui/material";
+>>>>>>> main
 import { styled } from "@mui/material/styles";
 
 export const ConfirmDialogContent = styled(DialogContent)`