+++ packages/admin/cms-admin/src/dam/MoveDamItemDialog/MoveDamItemDialog.tsx
@@ -1,4 +1,5 @@
 import { useApolloClient, useQuery } from "@apollo/client";
+<<<<<<< HEAD
 import { type FetchResult } from "@apollo/client/link/core";
 import { SaveButton } from "@comet/admin";
 import { Move, Reset } from "@comet/admin-icons";
@@ -10,6 +11,12 @@ import {
     DialogContent,
     DialogTitle,
 } from "@mui/material";
+=======
+import { FetchResult } from "@apollo/client/link/core";
+import { Button, SaveButton } from "@comet/admin";
+import { Move, Reset } from "@comet/admin-icons";
+import { Dialog, DialogActions, DialogContent, DialogTitle } from "@mui/material";
+>>>>>>> main
 import { styled } from "@mui/material/styles";
 import { useCallback } from "react";
 import { FormattedMessage } from "react-intl";
@@ -204,6 +211,7 @@ const MoveDamItemDialogInner = ({
             </DialogContent>
             <DialogActions>
                 <Button
+                    variant="textDark"
                     startIcon={<Reset />}
                     onClick={() => {
                         setSelectedId(undefined);