+++ packages/admin/cms-admin/src/pages/SaveConflictDialog.tsx
@@ -1,5 +1,6 @@
-import { Alert, messages } from "@comet/admin";
+import { Alert, Button, messages } from "@comet/admin";
 import { Clear, Delete, OpenNewTab } from "@comet/admin-icons";
+<<<<<<< HEAD
 import {
     Box,
     Button,
@@ -11,6 +12,9 @@ import {
     Stack,
     Typography,
 } from "@mui/material";
+=======
+import { Box, Dialog, DialogActions, DialogContent, DialogTitle, Stack, Typography } from "@mui/material";
+>>>>>>> main
 import { styled } from "@mui/material/styles";
 import { FormattedMessage } from "react-intl";
 
@@ -96,7 +100,7 @@ function SaveConflictDialog({ open, onClosePressed, onDiscardChangesPressed }: S
                 </Stack>
             </DialogContent>
             <DialogActions>
-                <Button onClick={onClosePressed} startIcon={<Clear />} color="info">
+                <Button variant="textDark" onClick={onClosePressed} startIcon={<Clear />}>
                     <FormattedMessage {...messages.close} />
                 </Button>
                 <DialogActionsSpacer />
@@ -106,8 +110,6 @@ function SaveConflictDialog({ open, onClosePressed, onDiscardChangesPressed }: S
                         onClosePressed();
                         onDiscardChangesPressed();
                     }}
-                    variant="contained"
-                    color="info"
                 >
                     <FormattedMessage id="comet.saveConflictDialog.actionButtons.discardChanges" defaultMessage="Discard your changes" />
                 </DiscardButton>
@@ -117,8 +119,6 @@ function SaveConflictDialog({ open, onClosePressed, onDiscardChangesPressed }: S
                         onClosePressed();
                         window.open(window.location.href, "_blank");
                     }}
-                    variant="contained"
-                    color="primary"
                 >
                     <FormattedMessage
                         id="comet.saveConflictDialog.actionButtons.openCurrentVersionInNewTab"