+++ packages/admin/cms-admin/src/pages/pagesPage/PagesPage.tsx
@@ -1,6 +1,7 @@
 import { useQuery } from "@apollo/client";
 import {
     Alert,
+    Button,
     Loading,
     LocalErrorScopeApolloContext,
     MainContent,
@@ -16,8 +17,13 @@ import {
     useStoredState,
 } from "@comet/admin";
 import { Add } from "@comet/admin-icons";
+<<<<<<< HEAD
 import { Box, Button, DialogContent, Divider, FormControlLabel, LinearProgress, Paper, Switch } from "@mui/material";
 import { type ComponentType, type ReactNode, useCallback, useMemo, useRef } from "react";
+=======
+import { Box, Divider, FormControlLabel, LinearProgress, Paper, Switch } from "@mui/material";
+import { ComponentType, ReactNode, useCallback, useEffect, useMemo, useRef } from "react";
+>>>>>>> main
 import { FormattedMessage, useIntl } from "react-intl";
 
 import { type ContentScopeInterface, createEditPageNode } from "../..";
@@ -138,8 +144,6 @@ export function PagesPage({
                             </ToolbarItem>
                             <ToolbarActions>
                                 <Button
-                                    variant="contained"
-                                    color="primary"
                                     startIcon={<Add />}
                                     onClick={() => {
                                         editDialogApi.openAddDialog();