+++ packages/admin/admin-icons/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/admin-icons
 
+<<<<<<< HEAD
 ## 8.0.0-beta.3
 
 ## 8.0.0-beta.2
@@ -33,6 +34,9 @@
 ### Minor Changes
 
 - 682a674: Add support for React 18
+=======
+## 7.17.0
+>>>>>>> main
 
 ## 7.16.0