+++ packages/admin/admin-rte/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/admin-rte
 
+<<<<<<< HEAD
 ## 8.0.0-beta.3
 
 ### Patch Changes
@@ -75,6 +76,14 @@
 - Updated dependencies [682a674]
     - @comet/admin@8.0.0-beta.0
     - @comet/admin-icons@8.0.0-beta.0
+=======
+## 7.17.0
+
+### Patch Changes
+
+-   @comet/admin@7.17.0
+-   @comet/admin-icons@7.17.0
+>>>>>>> main
 
 ## 7.16.0