+++ packages/admin/admin/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/admin
 
+<<<<<<< HEAD
 ## 8.0.0-beta.3
 
 ### Patch Changes
@@ -306,6 +307,14 @@
 - Updated dependencies [04e308a]
 - Updated dependencies [682a674]
     - @comet/admin-icons@8.0.0-beta.0
+=======
+## 7.17.0
+
+### Patch Changes
+
+-   @comet/admin-icons@7.17.0
+-   @comet/admin-theme@7.17.0
+>>>>>>> main
 
 ## 7.16.0